🎉 market营销服务初始化

+++ ahao-commerce-fulfill-api-old/src/main/java/com/ruyuan/eshop/fulfill/domain/request/ReceiveFulfillRequest.java
@@ -104,17 +104,17 @@ public class ReceiveFulfillRequest implements Serializable {
     /**
      * 付款总金额
      */
-    private Integer payAmount;
+    private BigDecimal payAmount;
 
     /**
      * 交易总金额
      */
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 
     /**
      * 运费
      */
-    private Integer deliveryAmount;
+    private BigDecimal deliveryAmount;
 
     /**
      * 用于模拟履约服务异常
