+++ ahao-commerce-order-api-old/src/main/java/com/ruyuan/eshop/order/domain/dto/OrderAmountDetailDTO.java
@@ -3,6 +3,7 @@ package com.ruyuan.eshop.order.domain.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * <p>
@@ -34,12 +35,12 @@ public class OrderAmountDetailDTO implements Serializable {
     /**
      * 销售数量
      */
-    private Integer saleQuantity;
+    private BigDecimal saleQuantity;
 
     /**
      * 销售单价
      */
-    private Integer salePrice;
+    private BigDecimal salePrice;
 
     /**
      * 收费类型
@@ -49,5 +50,5 @@ public class OrderAmountDetailDTO implements Serializable {
     /**
      * 收费金额
      */
-    private Integer amount;
+    private BigDecimal amount;
 }
