+++ ahao-commerce-order-api-old/src/main/java/com/ruyuan/eshop/order/domain/dto/ReleaseProductStockDTO.java
@@ -3,6 +3,7 @@ package com.ruyuan.eshop.order.domain.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -36,7 +37,7 @@ public class ReleaseProductStockDTO implements Serializable {
         /**
          * 销售数量
          */
-        private Integer saleQuantity;
+        private BigDecimal saleQuantity;
 
     }
