+++ ahao-commerce-order-service-old/src/main/java/com/ruyuan/eshop/order/domain/entity/AfterSaleRefundDO.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -68,7 +69,7 @@ public class AfterSaleRefundDO implements Serializable {
     /**
      * 退款金额
      */
-    private Integer refundAmount;
+    private BigDecimal refundAmount;
 
     /**
      * 退款支付时间
