+++ ahao-commerce-order-service-old/src/main/java/com/ruyuan/eshop/order/domain/entity/OrderInfoDO.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -88,12 +89,12 @@ public class OrderInfoDO implements Serializable {
     /**
      * 交易总金额（以分为单位存储）
      */
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 
     /**
      * 交易支付金额
      */
-    private Integer payAmount;
+    private BigDecimal payAmount;
 
     /**
      * 交易支付方式
