+++ ahao-commerce-product-api-old/src/main/java/com/ruyuan/eshop/product/domain/dto/ProductSkuDTO.java
@@ -3,6 +3,7 @@ package com.ruyuan.eshop.product.domain.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 
 /**
@@ -48,11 +49,11 @@ public class ProductSkuDTO implements Serializable {
     /**
      * 商品销售价格
      */
-    private Integer salePrice;
+    private BigDecimal salePrice;
 
     /**
      * 商品采购价格
      */
-    private Integer purchasePrice;
+    private BigDecimal purchasePrice;
 
 }
